+++ packages/oxlint-plugin-react-doctor/src/index.ts
@@ -4,13 +4,18 @@ export default reactDoctorPlugin;
 
 export {
   ALL_REACT_DOCTOR_RULE_KEYS,
+  ALL_REACT_DOCTOR_RULES,
+  BUILTIN_A11Y_RULES,
+  BUILTIN_REACT_RULES,
   FRAMEWORK_SPECIFIC_RULE_KEYS,
-  GLOBAL_REACT_DOCTOR_RULES,
   NEXTJS_RULES,
+  REACT_COMPILER_RULES,
   REACT_NATIVE_RULES,
+  RECOMMENDED_RULES,
   TANSTACK_QUERY_RULES,
   TANSTACK_START_RULES,
-} from "./rules-by-framework.js";
+  YOU_MIGHT_NOT_NEED_EFFECT_RULES,
+} from "./rules.js";
 
 export { MOTION_LIBRARY_PACKAGES } from "./plugin/constants/style.js";
